requirements/etl.txt
psycopg2-binary
redis
requests
pydantic
